+++ src/bzflag/SegmentedShotStrategy.cxx
@@ -336,7 +336,7 @@ void SegmentedShotStrategy::radarRender() const
   // Display leading lines
   if (length > 0) {
     const fvec3& vel = getPath().getVelocity();
-    const fvec3 dir = vel.normalize() * shotTailLength * length;
+    const fvec3 dir = vel.normalize() * shotTailLength * (float)length;
     glBegin(GL_LINES);
     glVertex2fv(orig);
     if (BZDBCache::leadingShotLine) {
